Fixed FOV Calc & Added Exceptions
Changed FOV calc to 1 - modifier.
Added exceptions to the ZoomAmount command

+++ src/main/java/com/nicoislost/owo/ZoomOConfigModel.java
@@ -8,13 +8,13 @@ import io.wispforest.owo.config.annotation.RangeConstraint;
 @Config(wrapperName = "ZoomOConfig", name = "Zoom-O-Config")
 public class ZoomOConfigModel {
     @RangeConstraint(min = 1, max = 100)
-    public double Zoom1=30;
+    public int Zoom1=30;
     public boolean Zoom1SmoothCamera = false;
     @RangeConstraint(min = 1, max = 100)
-    public double Zoom2=40;
+    public int Zoom2=40;
     public boolean Zoom2SmoothCamera = false;
     @RangeConstraint(min = 1, max = 100)
-    public double Zoom3=50;
+    public int Zoom3=50;
     public boolean Zoom3SmoothCamera = false;
 
 }
